+++ src/Voting.Lib.Cryptography.Mocks/DependencyInjection/ServiceCollectionExtensions.cs
@@ -4,6 +4,7 @@
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Voting.Lib.Cryptography;
 using Voting.Lib.Cryptography.Mocks;
+using Voting.Lib.Cryptography.Mocks.Configuration;
 
 namespace Microsoft.Extensions.DependencyInjection;
 
@@ -16,11 +17,15 @@ public static class ServiceCollectionExtensions
     /// Adds a PKCS#11 mock interface to perform cryptographic functions.
     /// </summary>
     /// <param name="services">The service collection.</param>
+    /// <param name="config">The crypto mock config.</param>
     /// <returns>The <see cref="IServiceCollection"/> instance.</returns>
-    public static IServiceCollection AddVotingLibCryptoProviderMock(this IServiceCollection services)
+    public static IServiceCollection AddVotingLibCryptoProviderMock(this IServiceCollection services, CryptoMockConfig? config = null)
     {
+        config ??= new CryptoMockConfig();
+
         return services
             .RemoveAll<ICryptoProvider>()
+            .AddSingleton(config)
             .AddSingleton<CryptoProviderMock>()
             .AddSingleton<ICryptoProvider>(sp => sp.GetRequiredService<CryptoProviderMock>());
     }